[mod] seperate melli into functional modules
