+++ security.py
@@ -1,0 +1,15 @@
+from functools import partial
+
+from jose import jwt
+from passlib.context import CryptContext
+
+# create secret key with: openssl rand -hex 32
+SECRET_KEY = "069d49a9c669ddc08f496352166b7b5d270ff64d3009fc297689aa8b0fb66d98"
+ALOGRITHM = "HS256"
+
+encode_jwt = partial(jwt.encode, key=SECRET_KEY, algorithm=ALOGRITHM)
+decode_jwt = partial(jwt.decode, key=SECRET_KEY, algorithms=[ALOGRITHM])
+
+_crypt_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
+verify_password = _crypt_context.verify
+hash_password = _crypt_context.hash
